+++ benchmark/camelyon/benchmarks.py
@@ -3,7 +3,9 @@ import multiprocessing
 import time
 from pathlib import Path
 
-from common.dataset_manager import creates_data_folders
+from classic_algos.nn import Weldon
+from common.dataset_manager import DATA_DIR
+from common.dataset_manager import creates_data_folder
 from common.dataset_manager import fetch_camelyon
 from common.dataset_manager import reset_data_folder
 from common.utils import parse_params
@@ -11,31 +13,58 @@ from common.utils import read_results
 from workflows import connectlib_fed_avg
 from workflows import torch_fed_avg
 
+from connectlib.index_generator import NpIndexGenerator
+
 RESULTS_FOLDER = Path(__file__).parent / "results"
 LOCAL_RESULTS_FILE = RESULTS_FOLDER / "results.json"
 
 
-def fed_avg(params: dict, train_folders, test_folders):
+def fed_avg(params: dict, train_folder: Path, test_folder: Path):
     """If remote, only running the benchmark on the connect platform (from the remote.yaml file) else running
     both connectlib and pure_torch fed avg strategy with the given parameters.
 
     Args:
         params (dict): Strategy and algo parameters.
+        train_folder (Path): Path to the data sample that will be used and duplicate for the benchmark.
+        test_folder (Path):  Path to the data sample that will be used and duplicate for the benchmark.
 
     Returns:
         dict: Results of both experiment with their computation time and the used parameters
     """
     exp_params = params.copy()
 
-    run_keys = ["n_rounds", "seed", "batch_size", "n_centers", "learning_rate", "n_local_steps", "num_workers", "mode"]
+    index_generator = NpIndexGenerator(
+        batch_size=exp_params["batch_size"], num_updates=exp_params["n_local_steps"], drop_last=True, shuffle=False
+    )
+
+    model = Weldon(
+        in_features=2048,
+        out_features=1,
+        n_extreme=10,
+        n_top=10,
+        n_bottom=10,
+    )
+
+    run_keys = [
+        "n_rounds",
+        "seed",
+        "n_centers",
+        "learning_rate",
+        "num_workers",
+        "nb_train_data_samples",
+        "nb_test_data_samples",
+    ]
 
     cl_start = time.time()
     cl_perf = connectlib_fed_avg(
-        train_folders=train_folders,
-        test_folders=test_folders,
+        train_folder=train_folder,
+        test_folder=test_folder,
         **{k: v for k, v in exp_params.items() if k in run_keys},
         credentials_path=exp_params["credentials"],
-        assets_keys_path=exp_params["assets_keys"],
+        asset_keys_path=exp_params["asset_keys"],
+        index_generator=index_generator,
+        model=model,
+        mode=exp_params["mode"],
     )
     cl_end = time.time()
     exp_params.update(
@@ -45,23 +74,22 @@ def fed_avg(params: dict, train_folders, test_folders):
         }
     )
 
-    if params["mode"] != "remote":
+    sa_start = time.time()
+    sa_perf = torch_fed_avg(
+        train_folder=train_folder,
+        test_folder=test_folder,
+        **{k: v for k, v in exp_params.items() if k in run_keys},
+        index_generator=index_generator,
+        model=model,
+    )
+    sa_end = time.time()
 
-        run_keys.pop()
-        sa_start = time.time()
-        sa_perf = torch_fed_avg(
-            train_folders=train_folders,
-            test_folders=test_folders,
-            **{k: v for k, v in exp_params.items() if k in run_keys},
-        )
-        sa_end = time.time()
-
-        exp_params.update(
-            {
-                "pure_torch_time": sa_end - sa_start,
-                "pure_torch_perf": sa_perf,
-            }
-        )
+    exp_params.update(
+        {
+            "pure_torch_time": sa_end - sa_start,
+            "pure_torch_perf": sa_perf,
+        }
+    )
 
     return {str(time.time()): exp_params}
 
@@ -83,23 +111,11 @@ def main():
 
     # Not use in remote, TODO: refactor at some point
     reset_data_folder()
-    train_folders = creates_data_folders(
-        n_centers=params.get("n_centers"),
-        sub_sampling=params.get("sub_sampling"),
-        data_samples_size=params.get("data_samples_size"),
-        batch_size=params.get("batch_size"),
-        kind="train",
-    )
-    test_folders = creates_data_folders(
-        n_centers=params.get("n_centers"),
-        sub_sampling=params.get("sub_sampling"),
-        data_samples_size=params.get("data_samples_size"),
-        batch_size=params.get("batch_size"),
-        kind="test",
-    )
+    train_folder = creates_data_folder(dest_folder=DATA_DIR / "train", index_path=DATA_DIR / "index.csv")
+    test_folder = creates_data_folder(dest_folder=DATA_DIR / "test", index_path=DATA_DIR / "index.csv")
 
     # Execute experiment
-    res = fed_avg(params, train_folders, test_folders)
+    res = fed_avg(params, train_folder, test_folder)
 
     if params["mode"] != "remote":
         # Update results